Added optics_utilities.py, and a generalazitation for the spacing. Need more testing. TODO add pytest example
