+++ solcore/optics/external_optics.py
@@ -2,35 +2,9 @@ from solcore.structure import Layer, Junction, TunnelJunction
 
 import numpy as np
 import types
+from optics_utilities import calculate_absorption, absorbed
 
 
-def calculate_absorption_rcwa(tmm_out):
-    all_z = tmm_out['position'] * 1e-9
-    all_abs = tmm_out['absorption'] / 1e-9
-
-    def diff_absorption(z):
-        idx = all_z.searchsorted(z)
-        idx = np.where(idx <= len(all_z) - 2, idx, len(all_z) - 2)
-        try:
-            z1 = all_z[idx]
-            z2 = all_z[idx + 1]
-
-            f = (z - z1) / (z2 - z1)
-
-            out = f * all_abs[:, idx] + (1 - f) * all_abs[:, idx + 1]
-
-        except IndexError:
-            out = all_abs[:, idx]
-
-        return out
-
-    all_absorbed = np.trapz(diff_absorption(all_z), all_z)
-
-    return diff_absorption, all_absorbed
-
-def absorbed(self, z):
-    out = self.diff_absorption(self.offset + z) * (z < self.width)
-    return out.T
 
 def solve_external_optics(solar_cell, options):
     """ Uses the reflection, transmission and absorption of a solar cell from
@@ -75,6 +49,7 @@ def solve_external_optics(solar_cell, options):
 
     # We calculate the total amount of light absorbed in the solar cell, 
     #integrating over its whole thickness with a step of 1 nm
+    # or using a user define mesh
     all_z = options.position
     all_absorbed = np.trapz(diff_absorption(all_z), all_z)
 
@@ -114,13 +89,13 @@ def solve_external_optics_matrix(solar_cell, options):
         raise err
 
     position = options.position * 1e9
-    delta = (position[1] - position[0])
-    absorption_power = absorbed_matrix / (delta)
+    # delta = (position[1] - position[0])
+    delta = 1./np.diff(position)
+    absorption_power = absorbed_matrix[:,1:] * delta
     output = {'position':   position, 
-              'absorption': None #np.zeros((wl.shape[0], len(dist)))}
+              'absorption': absorption_power,
              }
-    output['absorption'] = absorption_power
-    diff_absorption, all_absorbed = calculate_absorption_rcwa(output)
+    diff_absorption, all_absorbed = calculate_absorption(output)
 
     for j in range(len(solar_cell)):
         solar_cell[j].diff_absorption = diff_absorption
