+++ solcore/optics/rcwa.py
@@ -1,5 +1,6 @@
 from solcore.structure import Layer, Junction, TunnelJunction
 from solcore.absorption_calculator import calculate_rat_rcwa, calculate_absorption_profile_rcwa
+from optics_utilities import calculate_absorption, absorbed
 
 import numpy as np
 import types
@@ -57,7 +58,7 @@ def solve_rcwa(solar_cell, options):
                                             dist=position, theta=options.theta, phi=options.phi, pol=options.pol)
 
     # With all this information, we are ready to calculate the differential absorption function
-    diff_absorption, all_absorbed = calculate_absorption_rcwa(out)
+    diff_absorption, all_absorbed = calculate_absorption(out)
 
     # Each building block (layer or junction) needs to have access to the absorbed light in its region.
     # We update each object with that information.
@@ -70,31 +71,31 @@ def solve_rcwa(solar_cell, options):
     solar_cell.absorbed = all_absorbed * initial
 
 
-def absorbed(self, z):
-    out = self.diff_absorption(self.offset + z) * (z < self.width)
-    return out.T
+# def absorbed(self, z):
+    # out = self.diff_absorption(self.offset + z) * (z < self.width)
+    # return out.T
 
 
-def calculate_absorption_rcwa(tmm_out):
-    all_z = tmm_out['position'] * 1e-9
-    all_abs = tmm_out['absorption'] / 1e-9
+# def calculate_absorption_rcwa(tmm_out):
+    # all_z = tmm_out['position'] * 1e-9
+    # all_abs = tmm_out['absorption'] / 1e-9
 
-    def diff_absorption(z):
-        idx = all_z.searchsorted(z)
-        idx = np.where(idx <= len(all_z) - 2, idx, len(all_z) - 2)
-        try:
-            z1 = all_z[idx]
-            z2 = all_z[idx + 1]
+    # def diff_absorption(z):
+        # idx = all_z.searchsorted(z)
+        # idx = np.where(idx <= len(all_z) - 2, idx, len(all_z) - 2)
+        # try:
+            # z1 = all_z[idx]
+            # z2 = all_z[idx + 1]
 
-            f = (z - z1) / (z2 - z1)
+            # f = (z - z1) / (z2 - z1)
 
-            out = f * all_abs[:, idx] + (1 - f) * all_abs[:, idx + 1]
+            # out = f * all_abs[:, idx] + (1 - f) * all_abs[:, idx + 1]
 
-        except IndexError:
-            out = all_abs[:, idx]
+        # except IndexError:
+            # out = all_abs[:, idx]
 
-        return out
+        # return out
 
-    all_absorbed = np.trapz(diff_absorption(all_z), all_z)
+    # all_absorbed = np.trapz(diff_absorption(all_z), all_z)
 
-    return diff_absorption, all_absorbed
+    # return diff_absorption, all_absorbed
